Chopping trees
The tree_fall_small sound should alternate with the original tree falling sound when a tree is falling (Either this sound or the original will be played at random when a tree is falling)
The tree chopping sounds will constantly be played on loop when the area with the falling trees appear. When the background disappears, the far chopping will stop. When all of the trees disappear the close up chopping will stop