Fix city backgrounds and add new sounds (#4)
* Fixed city sidescrolling backgrounds

* Chopping trees

The tree_fall_small sound should alternate with the original tree falling sound when a tree is falling (Either this sound or the original will be played at random when a tree is falling)
The tree chopping sounds will constantly be played on loop when the area with the falling trees appear. When the background disappears, the far chopping will stop. When all of the trees disappear the close up chopping will stop

* Extra sounds

- Pageturn is the button press sound. Use it for the menu, settings, skip etc
- Bigbang will replace the current explosion sound. It will come after the narration of the sound unlike any other (also make sure all other sounds except the player sound are faded out before the explosion
- Wind-1 is for the starting cutscene, when the tree frame changes
- Distant warfare is the background sound throughout the war, instead of having music there. It should loop

---------

Co-authored-by: Erevoid <Erevoid@users.noreply.github.com>